+++ app/src/main/java/com/example/nexam/data/Exam.kt
@@ -19,7 +19,7 @@ data class Exam(
     @ColumnInfo(name = "difficulty")
     val difficulty: Int,
     @ColumnInfo(name = "remainingTime")
-    var remainingTime: Int,
+    var remainingTime: Long,
     @ColumnInfo(name = "additionalNotes")
     var additionalNotes: String,
     @ColumnInfo(name = "finished")
